refactor(StatsChart): extract to seperate file
